requirements.txt
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.2